+++ src/lib/constants.ts
@@ -81,3 +81,41 @@ export const DISPATCH_PRE_STORAGE_CATEGORIES = [
   "Consumption/Donation",
   "Transfer to other store",
 ] as const
+
+export const POTATO_VARIETIES = [
+  "Atlantic",
+  "Cardinal",
+  "Chipsona 1",
+  "Chipsona 2",
+  "Chipsona 3",
+  "Colomba",
+  "Desiree",
+  "Diamond",
+  "Suriya",
+  "FC - 11",
+  "FC - 12",
+  "FC - 5",
+  "Himalini",
+  "Fry Sona",
+  "KCM",
+  "K. Badshah",
+  "K. Chandramukhi",
+  "K. Jyoti",
+  "K. Pukhraj",
+  "Kuroda",
+  "Khyati",
+  "L.R",
+  "Lav Kar",
+  "Lima",
+  "Mohan",
+  "Pushkar",
+  "SU - Khyati",
+  "Super Six",
+] as const
+
+export type PotatoVariety = (typeof POTATO_VARIETIES)[number]
+
+export const POTATO_VARIETY_OPTIONS = POTATO_VARIETIES.map((value) => ({
+  id: value,
+  label: value,
+}))